+++ xamoomsdk/src/main/java/com/xamoom/android/xamoomsdk/Storage/OfflineStorageManager.java
@@ -10,14 +10,13 @@ package com.xamoom.android.xamoomsdk.Storage;
 
 import android.content.Context;
 import android.location.Location;
+import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
 
-import com.xamoom.android.xamoomsdk.APICallback;
 import com.xamoom.android.xamoomsdk.APIListCallback;
 import com.xamoom.android.xamoomsdk.Enums.ContentSortFlags;
-import com.xamoom.android.xamoomsdk.Enums.SpotFlags;
 import com.xamoom.android.xamoomsdk.Enums.SpotSortFlags;
-import com.xamoom.android.xamoomsdk.Offline.OfflineEnduserApi;
+import com.xamoom.android.xamoomsdk.Filter;
 import com.xamoom.android.xamoomsdk.Offline.OfflineEnduserApiHelper;
 import com.xamoom.android.xamoomsdk.Resource.Content;
 import com.xamoom.android.xamoomsdk.Resource.ContentBlock;
@@ -40,6 +39,7 @@ import java.io.IOException;
 import java.net.MalformedURLException;
 import java.net.URL;
 import java.util.ArrayList;
+import java.util.Date;
 import java.util.EnumSet;
 import java.util.List;
 
@@ -349,44 +349,22 @@ public class OfflineStorageManager {
   }
 
   /**
-   * Query content by tags.
+   * Returns contents from offline with the given filters.
+   * Used for calls with tags, names, dates and relatedSpots.
    *
-   * @param tags ArrayList<String> of tags.
+   * @param filter Filters for offline query
    * @param pageSize PageSize for paging. Limit 100.
    * @param cursor Cursor when paging. Null when first call.
    * @param sortFlags Sort query result.
    * @param callback Callback when finished.
    */
-  public void getContentByTags(List<String> tags, int pageSize, String cursor,
-                               EnumSet<ContentSortFlags> sortFlags,
-                               APIListCallback<List<Content>, List<Error>> callback) {
-
-    ArrayList<Content> contents = mContentDatabaseAdapter.getAllContents();
-    contents = OfflineEnduserApiHelper.getContentsWithTags(tags, contents);
-    OfflineEnduserApiHelper.PagedResult<Content> contentPagedResult =
-        OfflineEnduserApiHelper.pageResults(contents, pageSize, cursor);
-
-    sortContents(contents, sortFlags);
-
-    if (callback != null) {
-      callback.finished(contentPagedResult.getObjects(), contentPagedResult.getCursor(),
-          contentPagedResult.hasMore());
+  public void getContents(@NonNull Filter filter, int pageSize, @Nullable String cursor,
+                          EnumSet<ContentSortFlags> sortFlags,
+                          APIListCallback<List<Content>, List<Error>> callback) {
+    ArrayList<Content> contents = mContentDatabaseAdapter.getContents(filter);
+    if (filter != null && filter.getTags() != null) {
+      contents = OfflineEnduserApiHelper.getContentsWithTags(filter.getTags(), contents);
     }
-  }
-
-  /**
-   * Search content by name.
-   *
-   * @param name String to search for.
-   * @param pageSize PageSize for paging. Limit 100.
-   * @param cursor Cursor when paging. Null when first call.
-   * @param sortFlags Sort query result.
-   * @param callback Callback when finished.
-   */
-  public void searchContentsByName(String name, int pageSize, @Nullable String cursor,
-                                   EnumSet<ContentSortFlags> sortFlags,
-                                   APIListCallback<List<Content>, List<Error>> callback) {
-    ArrayList<Content> contents = mContentDatabaseAdapter.getContents(name);
 
     sortContents(contents, sortFlags);