+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # Changelog
 
-## [3.6.0](https://github.com/xamoom/xamoom-ios-sdk/compare/3.5.3...3.6.0) - 13.10.2017
+## [3.7.0](https://github.com/xamoom/xamoom-android-sdk/compare/3.6.0...3.7.0) - 30.10.2017
+
+- Added socialSharing url to content
+- Added eventPackage
+- Fixed phone links
+
+## [3.6.0](https://github.com/xamoom/xamoom-android-sdk/compare/3.5.3...3.6.0) - 13.10.2017
 
 - Added new content block: content list